backend_fastapi/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-dotenv>=1.0.1
